bugfix: fix render of empty grid
feat: refactor Palette component

feat: refactor file naming and structure
