+++ src/app/EditorCanvas/CanvasChunk/getSymbolBlock.ts
