+++ src/app/EditorControls/ColorsSelector/PaletteColorsSelector/PaletteColor/palette-color.module.css
