+++ src/app/EditorControls/ColorsSelector/PaletteColorsSelector/palette-colors-selector.module.css
