+++ src/app/EditorControls/ColorsSelector/colors-selector.module.css
