+++ src/app/EditorControls/ColorsSelector/useColorHotkeys.ts
