+++ src/app/EditorControls/SymbolsSelector/SymbolSelector/symbol-selector.module.css
