+++ src/app/EditorControls/SymbolsSelector/symbols-selector.module.css
